Now I've really, properly set things up for the tests

+++ spec/spec_helper.rb
@@ -7,13 +7,13 @@ Dir["./spec/support/**/*.rb"].each {|f| require f}
 
 RSpec.configure do |config|
   config.before :all do
-    FileUtils.mkdir '.awestruct'
+    FileUtils.mkdir('.awestruct') unless File.exists? '.awestruct'
+    $LOG = Logger.new('.awestruct/test.log') 
   end
   config.mock_framework = :rspec
   config.include NokogiriMatchers
   config.after :all do
-    FileUtils.rm_rf '.awestruct'
+    FileUtils.rm_rf '.awestruct' if File.exists? '.awestruct'
   end
 end
 
-$LOG = Logger.new(STDOUT)
